lib: Add a common mt_part header
A few boards and stage1 need to use mt_part functions from
the LK. However, they currently either re-define the mt_part*
functions they need with the address copied from the defconfig,
or (in better cases) directly use the symbol from autoconf.h
containing the address.

Either way, this is bad as we are defining mt_part* functions in
multiple places, both in stage1 and the board files.

Create a common mt_part header file to remove this duplication,
and to centralise the addresses to in the defconfig, as it should
be. Ironically, this involves essentially renaming the current
storage.h header to mt_part.h, and replacing it with an empty
placeholder for the upcoming storage lib.

For now, lib/mt_part.h will not be included in the board_ops.h
header, and boards will include it explicitly, this is because
we plan to move boards to use the new storage lib instead, which
provides a common abstraction between LK's that natively use the
new partition_* APIs and the legacy ones that use mt_part*.
